requirements.txt
pyserial==3.5
docopt==0.6.2